changing format of how data is layed out - searching by check id will just load all other data if it matches the correct type/target
