+++ src/views/agent/index.ts
@@ -1,7 +1,6 @@
 import NewAgent from "@/views/agent/NewAgent.vue";
 import Agent from "@/views/agent/Agent.vue";
 import AgentView from "@/views/agent/AgentView.vue";
-import AgentTarget from "@/views/agent/AgentTarget.vue"
 import check from "@/views/check";
 
 export default {
@@ -18,11 +17,6 @@ export default {
             path: '/agents/:agentId',
             name: 'agent',
             component: Agent,
-        },
-        {
-            path: '/agents/:agentId:/target/:targetHost:',
-            name: 'agentTarget',
-            component: AgentTarget,
         },
     ]
 }
